+++ pytket/docs/conf.py
@@ -38,9 +38,9 @@ copyright = "2022, Cambridge Quantum Computing Ltd"
 author = "Cambridge Quantum Computing Ltd"
 
 # The short X.Y version
-version = "1.7"
+version = "1.8"
 # The full version, including alpha/beta/rc tags
-release = "1.7.3"
+release = "1.8.1"
 
 
 # -- General configuration ---------------------------------------------------