+++ tket/src/Placement/GraphPlacement.cpp
@@ -47,6 +47,7 @@ GraphPlacement::default_pattern_weighting(const Circuit& circuit) const {
        gate_counter < this->maximum_pattern_gates_ && !frontier.slice->empty();
        i++) {
     for (const Vertex& vert : *frontier.slice) {
+      if (circuit.get_OpType_from_Vertex(vert) == OpType::Barrier) continue;
       EdgeVec q_out_edges =
           circuit.get_out_edges_of_type(vert, EdgeType::Quantum);
       unsigned n_q_edges = q_out_edges.size();
@@ -86,7 +87,8 @@ GraphPlacement::default_pattern_weighting(const Circuit& circuit) const {
       }
       if (n_q_edges > 2) {
         throw std::invalid_argument(
-            "Can only weight for Circuits with maximum two qubit gates.");
+            "Can only weight for Circuits with maximum two qubit quantum "
+            "gates.");
       }
     }
     frontier.next_slicefrontier();