+++ pytket/docs/changelog.rst
@@ -1,8 +1,16 @@
 Changelog
 =========
 
-1.8.0 (unreleased)
-------------------
+1.8.1 (November 2022)
+---------------------
+
+Fixes:
+
+* Incorrect qasm filtering.
+* Make graph placement work with multi-qubit barriers.
+
+1.8.0 (November 2022)
+---------------------
 
 Minor new features: